Fundamentals of Vector Use
Worked out the fundamentals of vector use. Some concerns of memory use, as each queue is initialized to the same capacity to avoid reallocation. It's all pointers to the Process objects and not the objects themselves, but at a million processes that's still 56MB for the queues alone, plus whatever memory the process objects themselves use and then the rest of the program.